fix(delete): fixed unit tests
